requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Werkzeug==2.3.7
cryptography==41.0.4
psycopg2-binary==2.9.7
python-socketio==5.8.0
python-engineio==4.7.1
requests==2.31.0
